requirements.txt
langchain
langchain-community
langchain_openai
unstructured
pdf2image
pytesseract
pillow
faiss-cpu
pdfminer
pi_heif
unstructured_inference